Updated authentication in the FAQ
Well, the FAQ still used the old authentication method with the `email` and `password` kwargs which could mislead someone reading the FAQ these days. I replaced these references with the base class of `fortnitepy.Auth` and a little comment telling them that they should use any type of authentication there. I think this could also be updated to not include the code snippet and instead refer to the `status` and `netcl` kwarg without the code