Fix startup errors not closing the process
Errors that occured on startup failed to close the process running do to it being raised in a task that wouldn't be awaited until the client had successfully started. In other words, the client would have to be started for the errors to be raised, which would of course not work because the startup would never finish.